# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "ActiveAlertsResponse": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/OfficialAlert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetadata"
          },
          "status": {
            "$ref": "#/components/schemas/AlertFeedStatus"
          }
        },
        "required": [
          "location",
          "alerts",
          "status",
          "meta"
        ],
        "title": "ActiveAlertsResponse",
        "type": "object"
      },
      "AlertFeedStatus": {
        "enum": [
          "available",
          "unavailable"
        ],
        "title": "AlertFeedStatus",
        "type": "string"
      },
      "CurrentConditions": {
        "additionalProperties": false,
        "properties": {
          "apparent_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apparent Temperature C"
          },
          "dewpoint_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dewpoint C"
          },
          "icon_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "pressure_hpa": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure Hpa"
          },
          "relative_humidity_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Humidity Pct"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "text_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Description"
          },
          "visibility_km": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility Km"
          },
          "wind_direction_deg": {
            "anyOf": [
              {
                "maximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Direction Deg"
          },
          "wind_gust_kph": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Gust Kph"
          },
          "wind_speed_kph": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Kph"
          }
        },
        "required": [
          "temperature_c",
          "apparent_temperature_c",
          "dewpoint_c",
          "text_description",
          "icon_url",
          "observed_at"
        ],
        "title": "CurrentConditions",
        "type": "object"
      },
      "CurrentWeatherResponse": {
        "additionalProperties": false,
        "properties": {
          "current": {
            "$ref": "#/components/schemas/CurrentConditions"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetadata"
          },
          "station": {
            "$ref": "#/components/schemas/ObservationStation"
          }
        },
        "required": [
          "location",
          "current",
          "station",
          "meta"
        ],
        "title": "CurrentWeatherResponse",
        "type": "object"
      },
      "DailyPeriod": {
        "additionalProperties": false,
        "properties": {
          "detailed_forecast": {
            "title": "Detailed Forecast",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "icon_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          },
          "is_daytime": {
            "title": "Is Daytime",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "minimum": 1.0,
            "title": "Number",
            "type": "integer"
          },
          "probability_of_precipitation_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability Of Precipitation Pct"
          },
          "short_forecast": {
            "title": "Short Forecast",
            "type": "string"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "wind_direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Direction"
          },
          "wind_speed_max_kph": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Max Kph"
          },
          "wind_speed_min_kph": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Min Kph"
          }
        },
        "required": [
          "number",
          "name",
          "start_time",
          "end_time",
          "is_daytime",
          "temperature_c",
          "wind_direction",
          "short_forecast",
          "detailed_forecast",
          "icon_url"
        ],
        "title": "DailyPeriod",
        "type": "object"
      },
      "DailyWeatherResponse": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetadata"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/DailyPeriod"
            },
            "title": "Periods",
            "type": "array"
          }
        },
        "required": [
          "location",
          "periods",
          "meta"
        ],
        "title": "DailyWeatherResponse",
        "type": "object"
      },
      "DataQuality": {
        "enum": [
          "verified",
          "partial",
          "unavailable"
        ],
        "title": "DataQuality",
        "type": "string"
      },
      "FreshnessStatus": {
        "enum": [
          "fresh",
          "delayed",
          "stale",
          "unavailable"
        ],
        "title": "FreshnessStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HourlyWeatherResponse": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetadata"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/HourlyPeriod"
            },
            "title": "Periods",
            "type": "array"
          }
        },
        "required": [
          "location",
          "periods",
          "meta"
        ],
        "title": "HourlyWeatherResponse",
        "type": "object"
      },
      "Location": {
        "additionalProperties": false,
        "properties": {
          "country": {
            "const": "US",
            "default": "US",
            "title": "Country",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/LocationKind"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "region",
          "latitude",
          "longitude",
          "timezone",
          "kind",
          "label"
        ],
        "title": "Location",
        "type": "object"
      },
      "LocationKind": {
        "enum": [
          "city",
          "zcta",
          "coordinate"
        ],
        "title": "LocationKind",
        "type": "string"
      },
      "LocationSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "query_kind": {
            "$ref": "#/components/schemas/QueryKind"
          },
          "region_limit_km": {
            "title": "Region Limit Km",
            "type": "number"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results",
          "query_kind",
          "region_limit_km"
        ],
        "title": "LocationSearchResponse",
        "type": "object"
      },
      "ObservationStation": {
        "additionalProperties": false,
        "properties": {
          "distance_km": {
            "minimum": 0.0,
            "title": "Distance Km",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "distance_km",
          "observed_at"
        ],
        "title": "ObservationStation",
        "type": "object"
      },
      "OfficialAlert": {
        "additionalProperties": false,
        "properties": {
          "area_description": {
            "title": "Area Description",
            "type": "string"
          },
          "certainty": {
            "title": "Certainty",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          },
          "issuing_office": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuing Office"
          },
          "message_type": {
            "title": "Message Type",
            "type": "string"
          },
          "onset_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onset At"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "sent_at": {
            "format": "date-time",
            "title": "Sent At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "urgency": {
            "title": "Urgency",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revision_id",
          "event",
          "headline",
          "description",
          "instruction",
          "area_description",
          "issuing_office",
          "sent_at",
          "effective_at",
          "onset_at",
          "expires_at",
          "ends_at",
          "severity",
          "certainty",
          "urgency",
          "status",
          "message_type",
          "response",
          "geometry"
        ],
        "title": "OfficialAlert",
        "type": "object"
      },
      "QueryKind": {
        "enum": [
          "city",
          "zip",
          "coordinate"
        ],
        "title": "QueryKind",
        "type": "string"
      },
...
      "ResponseMetadata": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "$ref": "#/components/schemas/SourceAttribution"
          },
          "confidence": {
            "title": "Confidence",
            "type": "null"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus"
          },
          "from_cache": {
            "default": false,
            "title": "From Cache",
            "type": "boolean"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "processed_at": {
            "format": "date-time",
            "title": "Processed At",
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "source_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Time"
          },
          "stale_fallback": {
            "default": false,
            "title": "Stale Fallback",
            "type": "boolean"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Units",
            "type": "object"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "source_time",
          "fetched_at",
          "processed_at",
          "valid_from",
          "valid_to",
          "units",
          "freshness",
          "quality",
          "pipeline_version"
        ],
        "title": "ResponseMetadata",
        "type": "object"
      },
      "SourceAttribution": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "default": "National Weather Service",
            "title": "Name",
            "type": "string"
          },
          "url": {
            "default": "https://www.weather.gov/",
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "title": "SourceAttribution",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Prairie Signal exposes attributed official observations, forecasts, and alerts. It does not provide radar, generated forecasts, or experimental hazard guidance.",
    "summary": "Official NWS weather data for the Lincoln service region",
    "title": "Prairie Signal",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alerts/active": {
      "get": {
        "operationId": "active_alerts_api_v1_alerts_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveAlertsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Alerts"
      }
    },
...
    "/api/v1/location/search": {
      "get": {
        "operationId": "search_locations_api_v1_location_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Locations"
      }
    },
...
    "/api/v1/weather/current": {
      "get": {
        "operationId": "current_weather_api_v1_weather_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current Weather"
      }
    },
    "/api/v1/weather/daily": {
      "get": {
        "operationId": "daily_weather_api_v1_weather_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyWeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily Weather"
      }
    },
    "/api/v1/weather/hourly": {
      "get": {
        "operationId": "hourly_weather_api_v1_weather_hourly_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 48,
              "maximum": 168,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HourlyWeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hourly Weather"
      }
    }
  }
}